Added rewind() method to DCD.
Improved error messaging in DCD().

Fixed bugs in dcdwriter.

Changed add() in Remarks to auto-linesplit if the input string is
longer than allowable in a PDB.

Changed interface to iterativeAlignment() so that it now returns a
Boost tuple containing the transformations for each input group as
well as the final dRMSD between average structures.

Improved alignment test to make sure that the returned composite
transform is correct.

Added tool to align frames of a DCD using iterativeAlignment.  Takes
two selections defining the subset to compute the alignment over and
the subset to apply the alignment too.  The defaults are CA's and all
non-solv atoms respectively.


svn path=/software/LOOS/trunk/; revision=576