Cleaned up MatrixWriter.hpp (it can now stand alone from LOOS...)
Changed invocationHeader() so that embedded newlines are replaced with
the string "\n".  This fixes a problems with multi-line selections
coming from a file...

Fixed bug in lexical scanner that allowed newlines to leak through to
stdout/cout.

rmsds now writes the # of atoms selected to cerr for conformation of
proper selection parsing.

svn path=/software/LOOS/trunk/; revision=681