Changed a few internal prototypes to be more prototype-ey and to use references to strings rather than copy the strings...  (dubious improvement)
Overloaded the parsing functions to work with either a string and a
range or just a string.

Added PDB::emptyString() function to check to see if a string contains
only spaces.

Changed PDB::parseConectRecord() to use fixed fields rather than
splitting based on spaces.  This should solve many of the problems
I've had with large PDBs.  Does not yet explicitly flag overflow
conditions...





svn path=/software/LOOS/trunk/; revision=1285