Changed the parseRangeList function.  It's been replaced with a family of templated functions:
   parseRange()       -- parses a single range
   parseRangeList()   -- parses a comma-separated list of ranges
   parseRangeList()   -- Take a vector of lists and parses 'em

The benefit is you can now get your vector of values back in whatever
type you want.  Since reverse ranges are supported (i.e. 5:0) and this
could be bad if you're using unsigned types, there is a minimal check
for this case.  You can still get nonsensical results if you are using
unsigned types and manually specify the step (i.e. 5:-1:3)...

A non-templated parseRangeList() is also included, returning a
vector<int> to be compatible with older versions of LOOS.



svn path=/software/LOOS/trunk/; revision=1677