Added a number of support/utility routines to support folding a generic SVD front-end into LOOS...
Notables include a routine to read a trajectory into a
vector<AtomicGroup>.

Added types RealMatrix and DoubleMatrix, which
are Math::Matrix<T, ColMajor> where T = float or double.

The svd function can take either a RealMatrix, a DoubleMatrix, or a
vector<AtomicGroup> (and an optional flag for whether or not to
automatically align the ensemble).  It will automatically subtract out
the average prior to computing the SVD.

Changed the iterativeAlignment functions to have a default threshold
of 1e-6...



svn path=/software/LOOS/trunk/; revision=1861