Cleaned up the SConstruct some more...
Added some hackage to support inserting a build code into the
invocationHeader output for the developer version...  This is
controlled by the REVISION variable in the SConstruct/custom.py.  If
it is set to 1 in your custom.py file, then the SConstruct will use
svnversion to try to figure out what the revision [ranges] are for
your source tree and insert that into invocationHeader.


svn path=/software/LOOS/trunk/; revision=1974