After talking to Alan, the build code now works as follows:
Setting the REVISION variable in custom.py to '' will cause the
SConstruct to call svnversion to figure out the current revision.  If
it's non-empty, then that will be used as the build revision.  The
default value is controlled by the loos_version variable set at the
start of the SConstruct.  This will need to be updated with each
release...




svn path=/software/LOOS/trunk/; revision=1975