Amber RST support finalized.
Magical incorporation of coords by createSystem(), when using a PRMTOP
and an INPCRD exists with the same name, has been removed.

Removed the obsolete Amber::readCoords()

Discovered an esoteric issue with Trajectory::cached_first that can
cause problems.  Granted, it's esoteric, but it suggested that this
mechanism was fragile.  Solving this required adding conditions on
functions that could perturb the cache status, i.e. the two seeks and
the rewind function.  These functions all now use NVI (or private
inheritance)...


svn path=/software/LOOS/trunk/; revision=1999