FINALLY, tracked down the problem with spurious complete rebuilds of LOOS.  It came from SCons generating implicit dependencies for the boost include files.  Mysteriously, one invocation adds a whole slew of BOOST headers as depencies, then the next invocation drops them. It may have something to do with order-dependence in generating the dependence signatures within SCons...I'm not sure...
Regardless, manually setting the include directories to the CPPFLAGS
solves this problem.  It's inelegant since it means we cannot rely on
BOOST to use the proper flags for different compilers and it also
means that changes to the BOOST and ATLAS headers will not be picked
up by SCons.  On the other hand, the frequent complete rebuilds was
driving me nuts, so this seems like a reasonable compromise for the
time being...


svn path=/software/LOOS/trunk/; revision=2464