reorganized matrix code.  Now support double and single precision matrices through simple overloading.
Permutation and overlap functions are now templatized and will work
with both single and double precision matrices.  Note that you cannot
mix and match, i.e. there is no C = A * B where A is single and B is
double.  Similarly, both sides of the overlap must be the same kind of
matrix...


svn path=/software/LOOS/trunk/; revision=3224