Added support for run-time selection of using either DSYGVX or DSYGVD as the eigensolver.  This is of little value at this point, but I wanted to have the code in here in case it's needed in the future.
Found a MAJOR bug in the diagnostic output.  The Msp matrix that was
written was not actually the Msp matrix.  This does NOT affect the
final output from the tool.  What it means, however, is that all
comparisons in Octave that were based on reading in the effective
matrices and doing the decomp there (i.e. either qz() or svd()) had
incorrect output.


svn path=/software/LOOS/trunk/; revision=3235