MAJOR fixes/changes...
Now uses an eigensolver for the generalized eigenproblem.  Also
mass-weights the eigenvectors and normalizes them.  Uses a cholesky
decomp to compute the square root of the effective mass matrix.

The intermediate matrices will only be written when debugging mode is
turned on (via command-line args).

Intermmediate matrices are all prefixed by the prefix...no, really!


svn path=/software/LOOS/trunk/; revision=3252