*MAJOR CHANGES*
Moved loadStructureWithCoords() and assignTrajectoryFrames() into LOOS
utils.

Longer options classes renamed to NOT have Options as a suffix.  I
decided this was pointless since it's almost always used as a
qualified name (i.e. OptionsFramework::BasicTrajectory).

Split trajectory options class into two:
      BasicTrajectory               Only supports --skip
      TrajectoryWithFrameIndices    Supports --skip and --range

Options classes that handle systems and trajectories now instantiate
an AtomicGroup (and pTraj) as part of the postConditions() functions.
This means the caller can reach in and copy the AtomicGroup if they
want to (or just use the one contained within the option object).

NOTE: The BasicTrajectory class will "prime" the
Trajectory::readFrame() iterator with the user-specified skip so that
it is ready to read from the correct frame.

Most of my tools use the TrajectoryWithFrameIndices class since they
were designed to support ranges from the outset.  atomic-rdf was
reverted to BasicTrajectory.



svn path=/software/LOOS/branches/Reorganization-2011/; revision=5948