Initial implementation.  Algorithm is a linear search with entries ordered by frequency (based on an scan of our local PSF files).
The actual data is hidden inside of a loos::internal class with the
user-level API being a function that has a static instance of the
deducer class.  I don't think this is sufficiently general enough to
require a loos-wide singleton (e.g. the RNG).  We'll see...



Also added a few missing headers to loos.hpp for completeness (they
were automatically included by headers that were already included, but
it looked odd not having them explicitly in there).


svn path=/software/LOOS/trunk/; revision=5986