added a (very long) fullhelp for order_params
I also made a small change to the code -- I noticed that when using the 1
residue/lipid format, the leading 2 or 3 in the carbon name was left on 
(eg the first carbon in the first chain is C22, and I deleteted the C but 
not the 2.  Now I delete both.

svn path=/software/LOOS/trunk/; revision=7650