Added better error messages when a prmtop file cannot be parsed.
Fixed a bug where the title string was not actually stored in _title
(this is a private variable however, so it shouldn't affect anything)

Added Amber::title() to return the private _title

Amber::verifyFormat() now throws a loos::ParseError upon failure.

Added support for having a title format of either 20a4 (the "official"
spec) or just "a" (used in the AMOEBA file from Alan).

Added support for varying numbers of pointers...  The first block
after the TITLE are the "pointers."  The official spec is for 32
pointers, however it turns out the the sample files used to write/test
amber.?pp only had 31.  So far, it seems we've been lucky in that
there are no reports of loos not working with prmtops, but conceivably
this could cause a parsing issue.  Ironically, the AMOEBA-derived
prmtop did have 32 pointers.

Now, we will read pointers until "%FLAG" is found, indicating the
start of the next block.

May want to think about switching to a line-oriented reader in the
future...


svn path=/software/LOOS/trunk/; revision=8739