Summary: Many fixes and changes to iterative alignment support in python
I forgot that Python doesn't really overload, so the Python iterativeAlignment() functions were broken...

iterativeAlignment() is now split into iterativeAlignEnsemble() that
takes a python list or an AtomicGroupVector() and iterativeAlignment()
that handles trajectories.  In both cases, the xformlist that is
returned is converted into a Python list containing XForms.

For the new iterativeAlignment(), if the passed traj is not a
loos.Trajectory, it will assume that whatever the object is has a
trajectory() method to access the underlying loos.Trajectory obj.

The framelist is now an optional named argument.  If it is not a
UIntVector(), then it as assumed to be iterable and one is created
from it for use with the C++ code

svn path=/software/LOOS/branches/PyAligned/; revision=14866