Summary: Added check for magic number when setting box with GCoord There's a "feature" where if you set a box, 	AtomicGroup g; 	AtomicGroup h;
	g.box(h.box());

and h's box is unset, you will end up with a box that has the magic
size and think it's set, but it's really not.  This makes it easy
to propagate bad box states.  Calls to isPeriodic() will return
true, even though the box size is a null-value.

The fix here is to check the GCoord used to set the box against
the magic value.  If it matches, then set the box size but change
the internal state to unset.

svn path=/software/LOOS/trunk/; revision=15188