Improved handling of documentation generation
Added support for automatically generating the documentation and handling pre-built
documentation (for distributions).

If a docs.built tag is found in the top-level dir, then auto-generation and cleaning of docs
is disabled.  Otherwise, docs are built as one would expect (i.e. it's generally rebuilt and
will be cleaned along with everything else).  The explicit docs target still remains.