Changed pre-built docs tag from docs.built to docs.prebuilt
The old loos would create a docs.built tag file, but "scons -c" wont clean that.  So, if you
had an older loos clone, pulled the latest docs changes in, and then ran scons, it would
complain because it couldn't generate the new docs.  So, we changed the tag file to
docs.prebuilt