Removed streamwrapper from Trajectory class and derived classes and added support functions
The StreamWrapper instance variable has been replaced with a Trajectory::pStream, which
is just a boost shared pointer to an istream instance.  When a Trajectory is
instantiated with a filename, the file is opened and a destroyable shared pointer to
that stream is created.  When a stream is used, a shared pointer with a noop deallocator
is used instead.

This breaks the interface for derived classes of Trajectory however.  Any reference to
ifs() should be replaced with just ifs.  If you need the underlying pointer, then
call ifs.get().

In order to provide some of the same functionality as StreamWrapper, Trajectory now has
a setInputStream() function that can be called with either a filename or an istream and
it will reset the ifs variable.  Note that this does NOT reset the Trajectory
object in any way.  It's still up to derived classes to correctly re-initialize if they
need to do so.