Updated range list parsing tools to accept wildcards.
Ranges can now include a wildcard for the endpoint, e.g. 10:* or 10:2:*
The first is 10 through the end and the second is 10 through the end taking every other one.

To be used properly, pass the expected max endpoint into the parseRange family of functions.
Nonsense can occur if a wildcard is used but the endpoint is unset (it defaults to 0).