Added top-level support for velocities.
Whether velocities are supported can be tested with Trajectory::hasVelocities().

The velocities of an AtomicGroup can be updated with Trajectory::updateGroupVelocities().
If the trajectory supports velocities, then those are used.  If not, then the coordinates
are copied into the velocities (with an appropriate conversion factor).

Subclasses need to override: hasVelocities(), velocityConversionFactor(), and updateGroupVelocitiesImpl().