Update lipid_unwrap.py
Added a top commnet for what the file does. Changed the a,b...f to one dictonary that was all the values of the tracker to the fix. added if __name__ in main . Spelling fixed(not well). Commneted on what i am doin in the trajectory making stuff. removed so silly thing about making the pdb . incresed the number of the new PBC Box