Update lipid_unwrap.py
Used the code given to me by Dr.Grossfeild That moved all no function calls to after the if name in main. editted if name in main to "__main__".  Removed old commnet that did not apply anymore around line 215, Added an arugment to not recenter the unwrapped trajectory. Made a 2D array instead of buibling one from scratch over time.